Move VSUM UI project to new framework UI namespace
