+++ bundles/framework/tools.vitruv.framework.vsum.ui/plugin.xml
@@ -4,14 +4,14 @@
    <extension
          point="org.eclipse.ui.newWizards">
       <category
-            id="tools.vitruv.framework.vsum.ui.Vitruv"
+            id="tools.vitruv.framework.ui.vsum.Vitruv"
             name="Vitruv">
       </category>
       <wizard
-            category="tools.vitruv.framework.vsum.ui.Vitruv"
-            class="tools.vitruv.framework.vsum.ui.wizard.CreateVsumWizard"
-            icon="src/tools/vitruv/framework/vsum/ui/resources/vitruvian.png"
-            id="tools.vitruv.framework.vsum.ui.newvitruvwizard"
+            category="tools.vitruv.framework.ui.vsum.Vitruv"
+            class="tools.vitruv.framework.ui.vsum.wizard.CreateVsumWizard"
+            icon="src/tools/vitruv/framework/ui/vsum/resources/vitruvian.png"
+            id="tools.vitruv.framework.ui.vsum.newvitruvwizard"
             name="Vitruv Project">
       </wizard>
    </extension>
@@ -21,21 +21,21 @@
             allPopups="false"
             locationURI="toolbar:org.eclipse.ui.main.toolbar">
          <toolbar
-               id="tools.vitruv.framework.vsum.ui.toolbar">
+               id="tools.vitruv.framework.ui.vsum.toolbar">
             <command
-                  commandId="tools.vitruv.framework.vsum.ui.vitruvwizardcommand"
-                  icon="src/tools/vitruv/framework/vsum/ui/resources/vitruvian.png"
+                  commandId="tools.vitruv.framework.ui.vsum.vitruvwizardcommand"
+                  icon="src/tools/vitruv/framework/ui/vsum/resources/vitruvian.png"
                   style="push">
             </command>
          </toolbar>
       </menuContribution>
    </extension>
    <extension
-         id="tools.vitruv.framework.vsum.ui.vitruvwizardcommand"
+         id="tools.vitruv.framework.ui.vsum.vitruvwizardcommand"
          point="org.eclipse.ui.commands">
       <command
-            defaultHandler="tools.vitruv.framework.vsum.ui.CreateVsumButtonHandler"
-            id="tools.vitruv.framework.vsum.ui.vitruvwizardcommand"
+            defaultHandler="tools.vitruv.framework.ui.vsum.CreateVsumButtonHandler"
+            id="tools.vitruv.framework.ui.vsum.vitruvwizardcommand"
             name="VitruvWizard">
       </command>
    </extension>
